queue/queue_backend/requirements.txt
asgiref==3.10.0
Django==4.2.25
django-cors-headers==4.9.0
django-filter==25.1
djangorestframework==3.16.1
importlib_metadata==8.7.0
Markdown==3.9
psycopg2-binary==2.9.11
python-decouple==3.8
sqlparse==0.5.3
typing_extensions==4.15.0
zipp==3.23.0
drf-spectacular
razorpay==2.0.0
qrcode[pil]==8.0
djangorestframework-simplejwt
gspread
django
google-auth==2.27.0
google-auth-oauthlib==1.2.0
pandas==2.1.4
openpyxl==3.1.2